+++ junit-jupiter-migration-support/src/main/java/org/junit/jupiter/migrationsupport/parameterized/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Extension to run JUnit 4 style parameterized tests in JUnit Jupiter.
+ */
+
+package org.junit.jupiter.migrationsupport.parameterized;
